refactor: rename `Iservice.ts` to `IService.ts`
