+++ src/services/speciality/ISpecialityService.ts
@@ -4,7 +4,7 @@ import type {
   SpecialtyUpdate,
 } from "@/models/types/Specialty.type.js";
 
-import type { IService } from "../Iservice.js";
+import type { IService } from "../IService.js";
 
 interface ISpecialityService extends IService<SpecialtyOutput> {
   create(data: SpecialtyInput): Promise<SpecialtyOutput>;
